♻️ refactor: reorganize package structure to follow Go best practices
- Move ECS package from pkg/ecs to internal/ecs to keep it internal to the project
- Extract generic slice helper functions to internal/utils/slice.go
- Update all import paths to reflect new internal package structure
- Follow Go idiomatic project layout with internal packages for non-exported code
