+++ cmd/deploy.go
@@ -10,7 +10,7 @@ import (
 	"syscall"
 
 	"github.com/spf13/cobra"
-	"runecs.io/v1/pkg/ecs"
+	"runecs.io/v1/internal/ecs"
 )
 
 func newDeployCommand() *cobra.Command {
