+++ cmd/list.go
@@ -11,7 +11,7 @@ import (
 	"github.com/charmbracelet/lipgloss/list"
 	"github.com/charmbracelet/lipgloss/table"
 	"github.com/spf13/cobra"
-	"runecs.io/v1/pkg/ecs"
+	"runecs.io/v1/internal/ecs"
 )
 
 func newListCommand() *cobra.Command {
