+++ cmd/revisions.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/charmbracelet/lipgloss"
 	"github.com/spf13/cobra"
-	"runecs.io/v1/pkg/ecs"
+	"runecs.io/v1/internal/ecs"
 )
 
 func newRevisionsCommand() *cobra.Command {
