+++ cmd/run.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/buildkite/shellwords"
 	"github.com/spf13/cobra"
-	"runecs.io/v1/pkg/ecs"
+	"runecs.io/v1/internal/ecs"
 )
 
 func newRunCommand() *cobra.Command {
