+++ internal/ecs/types.go
